Add a test to the web-test-runner
